making Multiples use long instead of int

+++ exercicio1/src/main/java/br/com/andre/exercicio1/Multiples.java
@@ -6,32 +6,32 @@ import java.util.Set;
 // Abstract Factory design pattern
 // 100% SOLID
 public final class Multiples implements OfRelationshipBetweenNumbers {
-    public OfRelationshipCalculator of(int operand) {
-        var baseOperands = new HashSet<Integer>();
+    public OfRelationshipCalculator of(long operand) {
+        var baseOperands = new HashSet<Long>();
         baseOperands.add(operand);
         return new MultiplesCalculator(baseOperands, new MultiplesOfAll());
     }
 
-    public OfRelationshipCalculator ofAll(int[] operands) {
-        Set<Integer> baseOperands = createOperandsSet(operands);
+    public OfRelationshipCalculator ofAll(long[] operands) {
+        Set<Long> baseOperands = createOperandsSet(operands);
         return ofAll(baseOperands);
     }
 
-    public OfRelationshipCalculator ofAll(Set<Integer> baseOperands) {
+    public OfRelationshipCalculator ofAll(Set<Long> baseOperands) {
         return new MultiplesCalculator(baseOperands, new MultiplesOfAll());
     }
 
-    public OfRelationshipCalculator ofAny(int[] operands) {
-        Set<Integer> baseOperands = createOperandsSet(operands);
+    public OfRelationshipCalculator ofAny(long[] operands) {
+        Set<Long> baseOperands = createOperandsSet(operands);
         return ofAny(baseOperands);
     }
 
-    public OfRelationshipCalculator ofAny(Set<Integer> operands) {
+    public OfRelationshipCalculator ofAny(Set<Long> operands) {
         return new MultiplesCalculator(operands, new MultiplesOfAny());
     }
 
-    private Set<Integer> createOperandsSet(int[] operands) {
-        var baseOperands = new HashSet<Integer>();
+    private Set<Long> createOperandsSet(long[] operands) {
+        var baseOperands = new HashSet<Long>();
 
         for (var i = 0; i < operands.length; i++) {
             baseOperands.add(operands[i]);
