+++ exercicio1/src/test/java/br/com/andre/exercicio1/Exercicio1_Tests.java
@@ -7,19 +7,19 @@ import org.junit.jupiter.api.Test;
 class Exercicio1_Tests {
     @Test
     void pergunta1() {
-        int respostaPergunta1 = new Multiples().ofAny(new int[] { 3, 5 }).until(999).sum();
+        long respostaPergunta1 = new Multiples().ofAny(new long[] { 3, 5 }).until(999).sum();
         assertEquals(233168, respostaPergunta1);
     }
 
     @Test
     void pergunta2() {
-        int respostaPergunta2 = new Multiples().ofAll(new int[] { 3, 5 }).until(999).sum();
+        long respostaPergunta2 = new Multiples().ofAll(new long[] { 3, 5 }).until(999).sum();
         assertEquals(33165, respostaPergunta2);
     }
 
     @Test
     void pergunta3() {
-        int respostaPergunta3 = new Multiples().ofAny(new int[] { 3, 5 }).until(999).sum();
+        long respostaPergunta3 = new Multiples().ofAny(new long[] { 3, 5 }).until(999).sum();
         respostaPergunta3 += new Multiples().of(7).until(999).sum();
         assertEquals(304239, respostaPergunta3);
     }
